requirements.txt
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
tqdm==4.66.2
joblib==1.3.2
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
scipy==1.12.0
matplotlib==3.8.3
pillow==10.2.0
python-dateutil==2.9.0.post0
ultralytics==8.2.6
roboflow==1.1.28
supervision==0.20.0
timm==1.0.7

